ref & feat: refactory folders of services - shadow entities; add redis and configuration to check status of user by token and session keycloak

+++ identity-service/src/main/java/com/example/feature/users/dto/UserUpdateRequest.java
@@ -19,6 +19,6 @@ public class UserUpdateRequest {
     private Integer gender;
 
     private String studentCode;
-    private Integer classId;
-    private Integer departmentId;
+    private Long classId;
+    private Long departmentId;
 }
